refactor(server): decompose AndroidGattServer (1003 loc) into 5 focused files
Extract state management, callback handling, setup logic, and extension
functions using the Class Decomposition Pattern (Facade/Registry/Emitter).

- AndroidGattServer: thin facade (286 lines)
- AndroidGattServerState: connection/subscription registry (204 lines)
- AndroidGattServerCallback: Binder callback + dispatch (444 lines)
- AndroidGattServerSetup: open, buildNativeService, notify delivery (183 lines)
- AndroidGattServerExtensions: property/permission mapping (27 lines)

All under the 300-line target except the callback (444 lines, inherently
verbose due to 9 required BluetoothGattServerCallback overrides).
